// src/types/database.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.5"
  }
  public: {
    Tables: {
// ...
      universities: {
        Row: {
          color: string
          email_domain: string | null
          id: string
          name: string
          short: string
        }
        Insert: {
          color: string
          email_domain?: string | null
          id: string
          name: string
          short: string
        }
        Update: {
          color?: string
          email_domain?: string | null
          id?: string
          name?: string
          short?: string
        }
        Relationships: []
      }
    }
// ...
  }
}